compiler warning fixed
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@1294 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_series_data.h
@@ -182,7 +182,7 @@ size_t QwtArraySeriesData<T>::size() const
 template <typename T>
 T QwtArraySeriesData<T>::sample( size_t i ) const
 {
-    return d_samples[i];
+    return d_samples[ static_cast<int>( i ) ];
 }
 
 //! Interface for iterating over an array of points
